refactor(architecture): break db/query circular dependency via QueryEngine trait
