+++ src/query/hybrid.rs
@@ -260,10 +260,10 @@ pub fn find_similar_as_of<G: GraphView + ?Sized>(
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::AletheiaDB;
     use crate::api::transaction::WriteOps;
     use crate::core::error::VectorError;
     use crate::core::property::PropertyMapBuilder;
-    use crate::db::AletheiaDB;
     use crate::index::vector::{DistanceMetric, HnswConfig};
 
     /// Helper to create a test database with vector indexing enabled.
