+++ src/query/mod.rs
@@ -114,4 +114,4 @@ pub use plan::{LogicalOp, LogicalPlan};
 pub use planner::{PhysicalPlan, QueryPlanner};
 pub use result::{EntityHistory, VersionDiff, VersionInfo, VersionSummary};
 pub use semantic_pathfinding::SemanticPathfinder;
-pub use traits::GraphView;
+pub use traits::{GraphView, QueryEngine};
