+++ src/query/planner/stats.rs
@@ -21,7 +21,7 @@
 //! 4. **Invalidation**: Schema changes (e.g., adding an index) invalidate the cache,
 //!    forcing a refresh on the next query.
 //!
-//! [`AletheiaDB::refresh_statistics`]: crate::db::AletheiaDB::refresh_statistics
+//! [`AletheiaDB::refresh_statistics`]: crate::AletheiaDB::refresh_statistics
 
 use std::sync::atomic::{AtomicBool, AtomicUsize, Ordering};
 
@@ -335,7 +335,7 @@ impl Statistics {
 
     /// Update statistics from storage.
     ///
-    /// This is called by [`AletheiaDB::refresh_statistics`](crate::db::AletheiaDB::refresh_statistics)
+    /// This is called by [`AletheiaDB::refresh_statistics`](crate::AletheiaDB::refresh_statistics)
     /// to populate the cache with fresh values from the storage engine.
     ///
     /// ## Examples
